Name bundler as Microsoft.DotNet.Build.Bundle
This commit makes two modifications:
* Name the bundler as Microsoft.DotNet.Build.Bundle.dll
* Reorganize the bundler code to separate files.
