+++ src/managed/Microsoft.DotNet.Build.Bundle/Microsoft.DotNet.Build.Bundle.csproj
