--- src/managed/tools/dir.proj
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
-
-
-  <Target Name="Build"
-          DependsOnTargets="GetLatestCommitHash">
-    <MakeDir Condition="!Exists('$(pOutDir)')" Directories="$(pOutDir)" />
-
-    <ItemGroup>
-      <Project Include="$(MSBuildThisFileDirectory)bundle/bundle.csproj" />
-    </ItemGroup>
-
-    <PropertyGroup>
-      <BuildArgs>$(NoRestoreArg) $(MSBuildPassThroughPropertyList)</BuildArgs>
-    </PropertyGroup>
-
-    <Exec Command="$(DotnetToolCommand) build $(BuildArgs) %(Project.Identity)" />
-  </Target>
-
-</Project>
