+++ src/test/Microsoft.DotNet.Build.Bundle.Tests/Microsoft.DotNet.Build.Bundle.Tests.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
-    <AssemblyName>BundleTests</AssemblyName>
-    <PackageId>BundleTests</PackageId>
+    <AssemblyName>Microsoft.DotNet.Build.Bundle.Tests</AssemblyName>
+    <PackageId>Microsoft.DotNet.Build.Bundle.Tests</PackageId>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <DefineConstants Condition="'$(OSGroup)' == 'Windows_NT'">$(DefineConstants);WINDOWS</DefineConstants>
   </PropertyGroup>
